+++ src/main/java/net/werdei/biome_replacer/MixinPlugin.java
@@ -16,11 +16,13 @@ public class MixinPlugin implements IMixinConfigPlugin
     private static final String TERRABLENDER_ID = "terrablender";
     private static final String BLUEPRINT_ID = "blueprint";
     private static final String BIOLITH_ID = "biolith";
+    private static final String LITHOSTITCHED_ID = "lithostitched";
 
     private String ownPackage;
     private boolean terrablenderInstalled;
     private boolean blueprintInstalled;
     private boolean biolithInstalled;
+    private boolean lithostitchedInstalled;
 
 
     @Override
@@ -30,12 +32,15 @@ public class MixinPlugin implements IMixinConfigPlugin
         terrablenderInstalled = Platform.isModLoaded(TERRABLENDER_ID);
         blueprintInstalled = Platform.isModLoaded(BLUEPRINT_ID);
         biolithInstalled = Platform.isModLoaded(BIOLITH_ID);
+        lithostitchedInstalled = Platform.isModLoaded(LITHOSTITCHED_ID);
         if (terrablenderInstalled)
             LOGGER.info(LOG_PREFIX, "TerraBlender detected, biome replacements will be injected into it");
         if (blueprintInstalled)
             LOGGER.info(LOG_PREFIX, "Blueprint detected, biome replacements will be injected into it");
         if (biolithInstalled)
             LOGGER.info(LOG_PREFIX, "Biolith detected, biome replacements will be injected into it");
+        if (lithostitchedInstalled)
+            LOGGER.info(LOG_PREFIX, "Lithostitched detected, biome replacements will be registered with it");
     }
 
     @Override
@@ -48,6 +53,8 @@ public class MixinPlugin implements IMixinConfigPlugin
             return blueprintInstalled;
         if (mixinShortName.startsWith(BIOLITH_ID))
             return biolithInstalled;
+        if (mixinShortName.startsWith(LITHOSTITCHED_ID))
+            return lithostitchedInstalled;
         return true;
     }